Create monitor package
Move subscriptions/notifications/spool modules/definitions under a new
package called "monitor". This allows avoiding cyclic imports in the
next commit introducing the "LIGHT_ASSERTS" variable, reduces the number
of top-level symbols and groups related modules.

This also renames the "kcidb-spool-wipe" executable to
"kcidb-monitor-spool-wipe", accordingly.
