+++ kcidb/monitor/__init__.py
@@ -1,53 +1,17 @@
-"""Kernel CI report subscriptions"""
+"""Kernel CI reporting - monitor"""
 
-import pkgutil
-import importlib
 import kcidb_io as io
 from kcidb import oo
-from kcidb.misc import Notification
+from kcidb.monitor import subscriptions, output, spool, misc
 
-
-def _load_match_map():
-    """
-    Load subscription match functions from kcidb.subscriptions.* modules.
-
-    Returns:
-        A dictionary of object list names and a list of tuples, each
-        containing a subscription name (subscription module name) and the
-        object type's matching function. The matching function accepts an
-        object of corresponding type, and returns an iterable of
-        kcidb.misc.NotificationMessage objects, or None, which is equivalent
-        to an empty iterable.
-    """
-    match_map = {
-        obj_list_name: []
-        for obj_list_name in io.schema.LATEST.tree
-        if obj_list_name
-    }
-
-    # For each kcidb.subscriptions.* module
-    for _, module_name, _ in pkgutil.walk_packages(path=__path__):
-        module = importlib.import_module(__name__ + "." + module_name)
-        # For each object list name
-        for obj_list_name in match_map:
-            # Record the subscription (module) name and match function, if any
-            assert obj_list_name.endswith("s")
-            function_name = "match_" + obj_list_name[:-1]
-            if function_name in module.__dict__:
-                function = module.__dict__[function_name]
-                assert callable(function)
-                match_map[obj_list_name].append((module_name, function))
-
-    return match_map
-
-
-# The default map of subscription matching functions:
-# a dictionary of object list names and a list of tuples, each containing a
-# subscription name (subscription module name) and the object type's matching
-# function. The matching function must accept an object of corresponding type,
-# and return an iterable of kcidb.misc.NotificationMessage objects, or None,
-# which is equivalent to an empty iterable.
-MATCH_MAP = _load_match_map()
+__all__ = [
+    "subscriptions",
+    "output",
+    "spool",
+    "misc",
+    "match_oo",
+    "match_new_io",
+]
 
 
 def match_oo(oo_data, match_map=None):
@@ -64,21 +28,21 @@ def match_oo(oo_data, match_map=None):
 
                     Each function must accept an object from the corresponding
                     object list in OO data, and return an iterable producing
-                    kcidb.misc.NotificationMessage objects, or None, which is
-                    equivalent to an empty iterable.
+                    kcidb.monitor.output.NotificationMessage objects, or None,
+                    which is equivalent to an empty iterable.
 
                     The default is a dictionary of matching functions from all
-                    kcidb.subscriptions.* modules, where each is called
-                    "match_<OBJ_NAME>", where "<OBJ_NAME>" is an object list
-                    name without the "s" ending.
+                    kcidb.monitor.subscriptions.* modules, where each is
+                    called "match_<OBJ_NAME>", where "<OBJ_NAME>" is an object
+                    list name without the "s" ending.
 
     Returns:
-        The list of notifications: kcidb.misc.Notification objects.
+        The list of notifications: kcidb.monitor.output.Notification objects.
     """
     assert oo.is_valid(oo_data)
 
     if match_map is None:
-        match_map = MATCH_MAP
+        match_map = subscriptions.MATCH_MAP
     notifications = []
     # For each object list and its subscription function list
     for obj_list_name, subscription_function_list in match_map.items():
@@ -94,8 +58,10 @@ def match_oo(oo_data, match_map=None):
             for obj in oo_data.get(obj_list_name, {}).values():
                 # For each matched notification message, if any
                 for message in function(obj) or []:
+                    assert isinstance(message, output.NotificationMessage)
                     notifications.append(
-                        Notification(obj_list_name, obj, subscription, message)
+                        output.Notification(obj_list_name, obj,
+                                            subscription, message)
                     )
     return notifications
 
@@ -122,8 +88,8 @@ def match_new_io(base_io, new_io, match_map=None, copy=True):
 
                     Each function must accept an object from the corresponding
                     object list in OO data, and return an iterable producing
-                    kcidb.misc.NotificationMessage objects, or None, which is
-                    equivalent to an empty iterable.
+                    kcidb.monitor.output.NotificationMessage objects, or None,
+                    which is equivalent to an empty iterable.
 
                     The default is a dictionary of matching functions from all
                     kcidb.subscriptions.* modules, where each is called
@@ -135,7 +101,7 @@ def match_new_io(base_io, new_io, match_map=None, copy=True):
                     Optional, default is True.
 
     Returns:
-        The list of notifications: kcidb.misc.Notification objects.
+        The list of notifications: kcidb.monitor.output.Notification objects.
     """
     assert io.schema.is_valid(base_io)
     assert io.schema.is_valid(new_io)
