+++ kcidb/monitor/test_spool.py
@@ -5,7 +5,7 @@ import dateutil.parser
 import kcidb
 
 
-class KCIDBSpoolMainFunctionsTestCase(kcidb.unittest.TestCase):
+class SpoolMainFunctionsTestCase(kcidb.unittest.TestCase):
     """Test case for main functions"""
 
     def test_wipe_main(self):
@@ -13,7 +13,8 @@ class KCIDBSpoolMainFunctionsTestCase(kcidb.unittest.TestCase):
         datetime_str = "2020-09-28 15:42:18.170439+03:00"
         datetime = dateutil.parser.isoparse(datetime_str)
         argv = [
-            "kcidb.spool.wipe_main", "-p", "project", "-c", "collection",
+            "kcidb.monitor.spool.wipe_main",
+            "-p", "project", "-c", "collection",
             datetime_str
         ]
         driver_source = textwrap.dedent(f"""
@@ -22,7 +23,8 @@ class KCIDBSpoolMainFunctionsTestCase(kcidb.unittest.TestCase):
             from dateutil.tz import tzoffset
             client = Mock()
             client.wipe = Mock()
-            with patch("kcidb.spool.Client", return_value=client) as Client:
+            with patch("kcidb.monitor.spool.Client",
+                       return_value=client) as Client:
                 status = function()
             Client.assert_called_once_with("collection", project="project")
             client.wipe.assert_called_once_with(until={repr(datetime)})
